requirements.txt
discord.py
wikipedia
urllib3
aiohttp
speedtest-cli
pyyaml
